tools/job-tailor/requirements.txt
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
beautifulsoup4>=4.12.0
click>=8.1.0
requests>=2.31.0
pyyaml>=6.0
